refactor: Change structure of serializer to engine->adapter one
