+++ filejacket/adapters/serializer/__init__.py
@@ -1,0 +1,50 @@
+"""
+Handler is a package for creating files in an object-oriented way,
+allowing extendability to any file system.
+
+Copyright (C) 2021 Gabriel Fontenelle Senno Silva
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+Should there be a need for contact the electronic mail
+`filejacket <at> gabrielfontenelle.com` can be used.
+"""
+
+from .pickle import PickleSerializer
+from .dictionary import (
+    FileDictionarySerializer,
+    FileWithContentDictionarySerializer,
+    FileDictionarySerializerReadonly
+)
+from .json import JSONSerializer
+from .json_simple import (
+    FileJsonSerializer,
+    FileWithContentJsonSerializer,
+    FileJsonSerializerReadonly
+)
+
+__all__ = [
+    # Pickle
+    "PickleSerializer",
+    # Full generic JSON
+    "JSONSerializer",
+    # Dictionary
+    "FileWithContentDictionarySerializer",
+    "FileDictionarySerializer",
+    "FileDictionarySerializerReadonly",
+    # Simplify JSON
+    "FileJsonSerializer",
+    "FileWithContentJsonSerializer",
+    "FileJsonSerializerReadonly"
+]
