+++ filejacket/adapters/serializer/pickle.py
@@ -1,0 +1,53 @@
+"""
+Handler is a package for creating files in an object-oriented way,
+allowing extendability to any file system.
+
+Copyright (C) 2021 Gabriel Fontenelle Senno Silva
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+Should there be a need for contact the electronic mail
+`filejacket <at> gabrielfontenelle.com` can be used.
+"""
+from __future__ import annotations
+
+from typing import Any
+
+from dill import dumps, loads, HIGHEST_PROTOCOL
+
+
+from filejacket.engines.serializer import Serializer
+
+__all__ = ["PickleSerializer"]
+
+
+class PickleSerializer(Serializer):
+    """
+    Class that allow handling of Serialization/Deserialization from object to pickle and from it to object.
+    This serializer don't support version.
+    """
+
+    @classmethod
+    def serialize(cls, source: Any) -> Any:
+        """
+        Method to serialize the input `source` using dill as extension to `pickle`.
+        """
+        return dumps(source, protocol=HIGHEST_PROTOCOL, recurse=True)
+
+    @classmethod
+    def deserialize(cls, source: Any) -> Any:
+        """
+        Method to deserialize the input `source` using dill as extension to `pickle`.
+        """
+        return loads(source, protocol=HIGHEST_PROTOCOL, recurse=True)
